Implemented binary solver
Binary solver compares the board as a binary number to presolved boards, a binary and operation between the solved board numbers and the current board will result in if the board has 3 in a row. This solution will only work when the solved numbers have been found for the board 3 by 3.